[.github] Move github-related consts and enums to shared - Allows eng/tools to use them
