+++ .github/workflows/test/set-status.test.js
@@ -3,7 +3,7 @@
 import { beforeEach, describe, expect, it } from "vitest";
 import { setStatusImpl } from "../src/set-status.js";
 
-import { CheckConclusion, CheckStatus, CommitStatusState } from "../src/github.js";
+import { CheckConclusion, CheckStatus, CommitStatusState } from "../../shared/src/github.js";
 import { createMockCore, createMockGithub } from "./mocks.js";
 
 describe("setStatusImpl", () => {
